+++ projects/demo-maps/src/app/app.module.ts
@@ -22,7 +22,7 @@ import { RouteMap3Component } from './route-components/route-example-events/rout
 import { RouteMap4Component } from './route-components/route-example-custom-layers/route-map4.component';
 import { RouteMap5Component } from './route-components/route-example-layout/route-map5.component';
 import { RouteMap6Component } from './route-components/route-example-layer-style/route-map6.component';
-import { RouteAboutComponent } from './route-components/route-about/route-about.component';
+import { RouteLicensesComponent } from './route-components/route-licenses/route-licenses.component';
 
 import { HttpClientModule } from '@angular/common/http';
 
@@ -37,7 +37,8 @@ import { HttpClientModule } from '@angular/common/http';
     RouteMap3Component,
     RouteMap4Component,
     RouteMap5Component,
-    RouteMap6Component
+    RouteMap6Component,
+    RouteLicensesComponent
   ],
   imports: [
     BrowserModule,