+++ projects/demo-maps/src/app/route-components/route-licenses/route-licenses.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { RouteLicensesComponent } from './route-licenses.component';
+
+describe('RouteLicensesComponent', () => {
+  let component: RouteLicensesComponent;
+  let fixture: ComponentFixture<RouteLicensesComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ RouteLicensesComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(RouteLicensesComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});