+++ projects/services-ogc/src/public-api.ts
@@ -8,4 +8,5 @@ export * from './lib/owc/types/owc-json';
 export * from './lib/owc/types/eoc-owc-json';
 export * from './lib/wps/wpsclient';
 export * from './lib/wps/wps_datatypes';
+export * from './lib/wps/cache';
 export * from './lib/wmts/wmtsclient.service';